requirements.txt
arrow==1.1.1
asgiref==3.4.1
binaryornot==0.4.4
boto3==1.18.35
botocore==1.21.35
cachetools==4.2.2
certifi==2021.5.30
cffi==1.14.6
chardet==4.0.0
charset-normalizer==2.0.3
click==8.0.1
colorama==0.4.4
cookiecutter==1.7.3
cryptography==3.4.7
dataclasses==0.6
defusedxml==0.7.1
Django==3.2.5
django-allauth==0.45.0
django-crispy-forms==1.12.0
django-tailwind==2.2.0
django-widgets-improved==1.5.0
djangorestframework==3.12.4
google-api-core==1.31.1
google-api-python-client==2.15.0
google-auth==1.34.0
google-auth-httplib2==0.1.0
googleapis-common-protos==1.53.0
httplib2==0.19.1
idna==3.2
isodate==0.6.0
Jinja2==3.0.1
jinja2-time==0.2.0
jmespath==0.10.0
MarkupSafe==2.0.1
numpy==1.21.1
oauthlib==3.1.1
packaging==21.0
pandas==1.3.1
Pillow==8.3.1
poyo==0.5.0
protobuf==3.17.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyJWT==2.1.0
pyparsing==2.4.7
python-dateutil==2.8.2
python-slugify==5.0.2
python3-openid==3.2.0
pytz==2021.1
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.7.2
s3transfer==0.5.0
six==1.16.0
sqlparse==0.4.1
text-unidecode==1.3
uritemplate==3.0.1
urllib3==1.26.6
wincertstore==0.2
